requirements_explain.txt
numpy
geopy
networkx
matplotlib
pandas
seaborn
torch_geometric
torch --index-url https://download.pytorch.org/whl/cpu